client-api: Merge http_headers into ruma-common
