+++ crates/ruma-common/Cargo.toml
@@ -26,7 +26,7 @@ default = ["client", "server"]
 client = []
 server = []
 
-api = ["dep:http", "dep:konst"]
+api = ["dep:date_header", "dep:http", "dep:konst"]
 js = ["getrandom?/js", "uuid?/js"]
 rand = ["dep:rand", "dep:getrandom", "dep:uuid"]
 
@@ -65,6 +65,7 @@ compat-optional = []
 as_variant = { workspace = true }
 base64 = { workspace = true }
 bytes = { workspace = true }
+date_header = { version = "1.0.5", optional = true }
 form_urlencoded = "1.0.0"
 getrandom = { version = "0.2.6", optional = true }
 http = { workspace = true, optional = true }
